+++ tests/python/pants_test/backend/project_info/tasks/BUILD
@@ -72,6 +72,7 @@ python_tests(
     'src/python/pants/build_graph',
     'tests/python/pants_test/subsystem:subsystem_utils',
     'tests/python/pants_test/tasks:task_test_base',
+    'tests/python/pants_test/backend/python/tasks:interpreter_cache_test_mixin',
   ]
 )
