+++ tests/python/pants_test/backend/project_info/tasks/test_export.py
@@ -26,11 +26,12 @@ from pants.base.exceptions import TaskError
 from pants.build_graph.register import build_file_aliases as register_core
 from pants.build_graph.resources import Resources
 from pants.build_graph.target import Target
+from pants_test.backend.python.tasks.interpreter_cache_test_mixin import InterpreterCacheTestMixin
 from pants_test.subsystem.subsystem_util import subsystem_instance
 from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
-class ExportTest(ConsoleTaskTestBase):
+class ExportTest(InterpreterCacheTestMixin, ConsoleTaskTestBase):
 
   @classmethod
   def task_type(cls):
